casuarina: rro_overlays: Drop config_wifi5ghzSupport overlay
This is false; according to stock it is supported. But during network
searches, the device can only find 2.4GHz networks, and no 5GHz networks
are visible.

To test it further, when the device is connected to a router that provides
a combined 2.4GHz + 5GHz signal, the device automatically defaults back to
2.4GHz, means there is no 5GHz signal at all.

Drop this overlay to avoid any future misclaiming.

Change-Id: Id8c400c94a9f422d7c7ec2afdda04cc2d5a77a0b